requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
sqlparse==0.4.3
